replibert/pyproject.toml
[tool.poetry]
name = "replibert"
version = "0.1.0"
description = ""
authors = ["Yannik Queisler <dev.yannik.queisler@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1"
matplotlib = "^3.9"
torch = { version = "^2.5", optional = true }
pyyaml = "^6.0"
datasets = "^3.0"
nltk = "^3.9"
transformers = "^4.46.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"

[tool.poetry.group.cpu]
optional = true

[tool.poetry.group.cpu.dependencies]
torch = { version = "^2.5", source = "torch_cpu", markers = "extra=='cpu' and extra!='gpu'" }

[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
torch = { version = "^2.5", source = "torch_cuda", markers = "extra=='gpu' and extra!='cpu'" }

[tool.poetry.extras]
cpu = ["torch_cpu"]
gpu = ["torch_cuda"]

[[tool.poetry.source]]
name = "torch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[[tool.poetry.source]]
name = "torch_cuda"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
